% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots.samples.R
\name{boots.samples}
\alias{boots.samples}
\title{Bootstrap Samples}
\usage{
boots.samples(dat, sub.id, B)
}
\arguments{
\item{dat}{data frame or data table of our interest}

\item{sub.id}{String vector, the column name of the subject ID's.}

\item{B}{Natural number. Number of bootstrap samples to create.}
}
\value{
A list of length B, of matrices with index and subjects
}
\description{
Bootstrap Samples
}
\examples{
set.seed(1249)
subjects<-c("Sarah","John","Beth","Anna","Chris","Blake")
example.subject<-sample(subjects, size = 50,replace = TRUE)#' example.dat<-data.frame("Y"=rnorm(n=length(example.subject)),
                        "X1"=rpois(n=length(example.subject), lambda = 3),
                        "X2"=rnorm(n=length(example.subject)),
                        "X3"=rbeta(n=length(example.subject), shape1 = 3, shape2 = 0.5),
                        "subjects"=example.subject)
output<-boots.samples(dat=example.dat,sub.id = "subjects",B=4) #create 4 bootstrap samples
output[[1]] #This is the first sample. A data table with two columns, an index column and the new subject names
}
