% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_lmer.R
\name{boots_lmer}
\alias{boots_lmer}
\title{Bootstrap Sampled Linear Mixed Effects Models (LMERs)}
\usage{
boots_lmer(y, X, dat, boots.samples.list, use.formula = NULL, num_workers = 2L)
}
\arguments{
\item{y}{character string. Name of column in each data.table of boots.samples.list, to be used as the outcome in LMER.}

\item{X}{a vector of character string.  Names of the of covariates to fit in LMER, as they appear on each dataset of boots.samples.list.}

\item{dat}{a data frame or data table, which contains y, X. Also should contain the subject column which was used to generate boots.samples.list.}

\item{boots.samples.list}{boots.samples output.}

\item{use.formula}{a vector of class formula. This is to be used in lmer fit. If not provided, then a basic additive fixed effects model with random effect of (1|no.repeat.sub.id) will be fit.}

\item{num_workers}{integer. A number of cores to use for parallel computing.}
}
\value{
A list of length B, of linear mixed effects model fits.
}
\description{
The function returns a fitted LMER model using bootstrap sample data, and returns inference about the fits.
}
\examples{
set.seed(1249)
subjects<-c("Sarah","John","Beth","Anna","Chris","Blake")
example.subject<-sample(subjects, size = 50,replace = TRUE)
example.dat<-data.frame("Y"=rnorm(n=length(example.subject)),
                        "X1"=rpois(n=length(example.subject), lambda = 3),
                        "X2"=rnorm(n=length(example.subject)),
                        "X3"=rbeta(n=length(example.subject), shape1 = 3, shape2 = 0.5),
                        "subjects"=example.subject)
output<-boots_samples(dat=example.dat,sub.id = "subjects",B=4) #create 4 bootstrap samples
lmer.out<-boots_lmer(y="Y", X=c("X1","X2","X3"), dat=example.dat, boots.samples.list = output)
}
