% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_samples.R
\name{boots_samples}
\alias{boots_samples}
\title{Bootstrap Samples}
\usage{
boots_samples(dat, sub.id, B)
}
\arguments{
\item{dat}{data frame or data table of our interest}

\item{sub.id}{String vector, the column name of the subject ID's.}

\item{B}{Natural number. Number of bootstrap samples to create.}
}
\value{
A list of data.table of length B. Each data table contains \code{index} column and \code{no.repeat.id} column.
The \code{index} column shows the row number of the original dataset that were selected as bootstrap samples.
Therefore, in order to get the bootstrap data, the user has to use the index column as row number in the original dataset.
\code{no.repeat.id} column shows the 'sub.id' concatenated by trailing '__#' to represent subjects repeated multiple times.
}
\description{
Bootstrap Samples
}
\examples{
set.seed(1249)
subjects<-c("Sarah","John","Beth","Anna","Chris","Blake")
example.subject<-sample(subjects, size = 50,replace = TRUE)
example.dat<-data.frame("Y"=rnorm(n=length(example.subject)),
                        "X1"=rpois(n=length(example.subject), lambda = 3),
                        "X2"=rnorm(n=length(example.subject)),
                        "X3"=rbeta(n=length(example.subject), shape1 = 3, shape2 = 0.5),
                        "subjects"=example.subject)
output<-boots_samples(dat=example.dat,sub.id = "subjects",B=4) #create 4 bootstrap samples
head(output[[1]]) #This is the first outcome.
bootstrap_dat1<-example.dat[output[[1]]$index,] #this is the bootstrap sampled dataset from the first bootstrap run.
bootstrap_dat1$no.repeat.id<-output[[1]]$no.repeat.id #this code adds the correctly named subjects to the bootstrap data so that when the subject is used for LMER fit, it properly accounts for the subjects.
}
