% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Generate Bootstrap Samples and Statistics}
\usage{
bootstrap(data, statistic, nboot = 1000, return_samples = TRUE)
}
\arguments{
\item{data}{A vector, matrix, or data frame for bootstrapping.}

\item{statistic}{A user-supplied function that computes a statistic on the data.}

\item{nboot}{Number of bootstrap replicates (default: 1000).}

\item{return_samples}{Logical; whether to return bootstrap samples (default: TRUE).}
}
\value{
A list of class \code{bootstrap_results}:
\item{thetahat}{The observed statistic.}
\item{thetastar}{A vector of bootstrap replicates.}
\item{bsamples}{A list of bootstrap samples.}
\item{data}{The original data.}
\item{statistic}{The user-supplied statistic function.}
}
\description{
This function generates bootstrap replicates and provides key statistics.
}
\examples{
# Here is an example of bootstrapping the slope coefficient in a linear reg model
beta <- function(dat) {
mod <- lm(Sepal.Length ~ Sepal.Width, data = dat)
bb <- mod$coefficients[2]
return(bb)
}
set.seed(124)
bootstrap(data=iris, statistic=beta, nboot=2, return_samples=FALSE)
}
