% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_t.R
\name{bootstrap_t}
\alias{bootstrap_t}
\title{Compute Bootstrap Confidence Intervals}
\usage{
bootstrap_t(
  boot_obj,
  sdfun = NULL,
  Bsd = 25,
  method = c("boott-plugin", "boott-nested", "bootsym-plugin", "bootsym-nested"),
  alpha = 0.05,
  parallel = FALSE,
  cores = NULL,
  ...
)
}
\arguments{
\item{boot_obj}{An object created by the \code{bootstrap} function.}

\item{sdfun}{Optional; function to compute standard errors.}

\item{Bsd}{Number of bootstrap samples to estimate the standard error (default: 25).}

\item{method}{Confidence interval method: "boott-plugin", "boott-nested", "bootsym-plugin" or "bootsym-nested".}

\item{alpha}{Confidence level (default: 0.05).}

\item{parallel}{A logical value indicating whether parallel processing should be used in the
bootstrap computation (default is \code{FALSE}).}

\item{cores}{The number of cores used for parallel processing if \code{parallel} is set to \code{TRUE}.
This parameter is ignored if \code{parallel} is \code{FALSE}.}

\item{...}{Additional arguments to be passed to the \code{sdfun} function.}
}
\value{
A list containing the confidence intervals, bias, and standard errors.
}
\description{
Computes confidence intervals using:
\code{boott-plugin}:bootstrap-t, equal-tailed with plugin standard error,
\code{boott-nested}:bootstrap-t, equal-tailed with nested bootstrap (this is when there is no plugin se),
\code{bootsym-plugin}:bootstrap-t, symmetric with plugin standard error,
\code{bootsym-nested}:bootstrap-t, symmetric with nested bootstrap (this is when there is no plugin se),
}
