% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example.dat}
\alias{example.dat}
\title{example.dat data}
\format{
An object of class \code{data.frame} with 50 rows and 5 columns.
}
\usage{
data(example.dat)
}
\arguments{
\item{example.dat}{simulated data}
}
\description{
This data was simulated for vignettes.
}
\examples{
head(data(example.dat))
}
\keyword{datasets}
