% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lmer_out}
\alias{lmer_out}
\title{boots_lmer output in vingette}
\format{
An object of class \code{boots_output} of length 4.
}
\usage{
data(lmer_out)
}
\arguments{
\item{lmer_out}{Simulated outpupt}
}
\description{
boots_lmer output in vingette
}
\examples{
head(data(lmer_out))
}
\keyword{datasets}
