% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output}
\alias{output}
\title{boots_samples output in vingette}
\format{
An object of class \code{list} of length 5000.
}
\usage{
data(output)
}
\arguments{
\item{output}{Simulated outpupt}
}
\description{
boots_samples output in vingette
}
\examples{
head(data(output))
}
\keyword{datasets}
