% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.boots.lmer.pkg_bootsout.R
\name{plot.boots.lmer.pkg_bootsout}
\alias{plot.boots.lmer.pkg_bootsout}
\title{Plotting the boots.lmer output}
\usage{
\method{plot}{boots.lmer.pkg_bootsout}(x)
}
\arguments{
\item{x}{An object of class \code{boots.lmer.pkg_bootsout}.}

\item{y}{Ignored.}

\item{...}{Further arguments passed to
\code{\link[graphics:plot.window]{plot.window}}.}
}
\description{
Plotting the boots.lmer output
}
