% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_lmer.R, R/plot.boots_output.R,
%   R/summary.boots_output.R
\name{boots_lmer}
\alias{boots_lmer}
\alias{plot.boots_output}
\alias{summary.boots_output}
\title{Bootstrap Sampled Linear Mixed Effects Models (LMERs)}
\usage{
boots_lmer(y, X, dat, boots.samples.list, use.formula = NULL, num_workers = 2L)

\method{plot}{boots_output}(x, y, ...)

\method{summary}{boots_output}(object, ...)
}
\arguments{
\item{y}{Ignored.}

\item{X}{a vector of character string.  Names of the of covariates to fit in LMER, as they appear on each dataset of boots.samples.list.}

\item{dat}{a data frame or data table, which contains y, X. Also should contain the subject column which was used to generate boots.samples.list.}

\item{boots.samples.list}{\code{\link[boots.lmer:boots_samples]{boots_samples}} output.}

\item{use.formula}{a vector of class formula. This is to be used in lmer fit. If not provided, then a basic additive fixed effects model with random effect of (1|no.repeat.sub.id) will be fit.}

\item{num_workers}{integer. A number of cores to use for parallel computing.}

\item{x}{An object of class \code{boots_output}.}

\item{...}{Further arguments passed to}
}
\value{
A list of length 4.

\code{Error.Index} is a vector of logicals. The corresponding \code{boots.samples.list} input where \code{Error.Index ==TRUE} had an error fitting the lme4::lmer function.

\code{est.betas} is a list of length same as the \code{X} input. Each list shows estimated coefficient of the corresponding covariates using bootstrap samples.

\code{Estimates} is a data frame of the bootstrap output. It contains "Mean", "SD", "2.5\%", and "97.5\%" columns.
The "Mean" column is calculated by using \code{\link[base:mean]{mean}} function on the \code{est.betas}.
The "SD" column is calculated by using \code{\link[stats:sd]{sd}} function on the \code{est.betas}.
The "2.5\%" and "97.5\%" confidence intervals are calculated using \code{Mean + c(-1, 1)*qnorm(0.975)*SD}.

A boxplot of the estimated effects of covarites from bootstrap samples.

Inference about the bootstrap sampled LMER fits.
}
\description{
The function returns a fitted LMER model using bootstrap sample data, and returns inference about the fits.

This is a generic function returns to an object class \code{boots_output} from \code{\link[boots.lmer:boots_lmer]{boots_lmer}}.

This is a generic function returns to an object class \code{boots_output} from \code{\link[boots.lmer:boots_lmer]{boots_lmer}}.
}
\examples{
set.seed(1249)
subjects<-c("Sarah","John","Beth","Anna","Chris","Blake")
example.subject<-sample(subjects, size = 50,replace = TRUE)
example.dat<-data.frame("Y"=rnorm(n=length(example.subject)),
                        "X1"=rpois(n=length(example.subject), lambda = 3),
                        "X2"=rnorm(n=length(example.subject)),
                        "X3"=rbeta(n=length(example.subject), shape1 = 3, shape2 = 0.5),
                        "subjects"=example.subject)
output<-boots_samples(dat=example.dat,sub.id = "subjects",B=4) #create 4 bootstrap samples
lmer.out<-boots_lmer(y="Y", X=c("X1","X2","X3"), dat=example.dat, boots.samples.list = output)

##Print the summary
summary(lmer.out)

##Boxplot of the estimated coefficients
plot(lmer.out)

}
